add comment to files

+++ wp-content/themes/twenty-twenty-child/includes/wp-metabox.php
@@ -22,18 +22,21 @@ function products_add_fields(){
     
  
     // Get Value of Fields From Database
-    $description_textareafield = get_post_meta( $post->ID, '_description_textarea_field', true);
-    $price_numberfield = get_post_meta( $post->ID, '_price_number', true);
-    $sale_price_numberfield = get_post_meta( $post->ID, '_sale_price_number', true);
-    $youtube_video_textfield = get_post_meta( $post->ID, '_youtube_video_text_field', true);
-    $checkbox_stored_meta = get_post_meta( $post->ID );
 
-      // Get WordPress' media upload URL
+    //description field
+    $description_textareafield = get_post_meta( $post->ID, '_description_textarea_field', true);
+    //Price field
+    $price_numberfield = get_post_meta( $post->ID, '_price_number', true);
+    //Sale Price field
+    $sale_price_numberfield = get_post_meta( $post->ID, '_sale_price_number', true);
+    //YouTube video field
+    $youtube_video_textfield = get_post_meta( $post->ID, '_youtube_video_text_field', true);
+    //Is on sale?
+    $checkbox_stored_meta = get_post_meta( $post->ID );
+    // Get WordPress' media upload URL
     $upload_link = esc_url( get_upload_iframe_src() );
-    
     // See if there's a media id already saved as post meta
     $your_img_id = get_post_meta( get_the_ID(), '_your_img_id', true );
-    
     // Get the image src
     $your_img_src = wp_get_attachment_image_src( $your_img_id, 'full' );
     
@@ -53,15 +56,18 @@ function products_add_fields(){
 
 
   <div class="flex flex-wrap -mx-3 my-6">
+
+  <!-- Price -->
     <div class="w-full md:w-1/2 px-3">
       <label class="block uppercase tracking-wide text-gray-700 text-xs font-bold mb-2" for="grid-last-name">
         <?php echo __('Price', 'twentytwenty_child') ?>
       </label>
       <input
         class="appearance-none block w-full bg-gray-200 text-gray-700 border border-gray-200 rounded py-3 px-4 leading-tight focus:outline-none focus:bg-white focus:border-gray-500"
-        type="number" min="0" name="_price_number" value="<?php echo $price_numberfield; ?>" placeholder="Price">
+        type="number" min="0" name="_price_number" value="<?php echo $price_numberfield; ?>" placeholder="<?php echo __('Price', 'twentytwenty_child') ?>">
     </div>
 
+    <!-- Sale price -->
     <div class="w-full md:w-1/2 px-3 mb-6 md:mb-0">
       <label class="block uppercase tracking-wide text-gray-700 text-xs font-bold mb-2" for="grid-first-name">
         <?php echo __('Sale price', 'twentytwenty_child') ?>
@@ -69,22 +75,23 @@ function products_add_fields(){
       <input
         class="appearance-none block w-full bg-gray-200 text-gray-700 border border-gray-200 rounded py-3 px-4 leading-tight focus:outline-none focus:bg-white focus:border-gray-500"
         type="number" min="0" name="_sale_price_number" value="<?php echo $sale_price_numberfield; ?>"
-        placeholder="Sale price">
+        placeholder="<?php echo __('Sale price', 'twentytwenty_child') ?>">
     </div>
 
+    <!-- Is on sale? -->
     <div class="w-full md:w-1/2 px-3 mt-6">
       <div class="md:w-1/3"></div> <label class="md:w-2/3 block text-gray-500 font-bold">
         <input class="mr-2 leading-tight" type="checkbox" name="_on_sale_checkbox_field" value="yes"
           <?php if ( isset ( $checkbox_stored_meta ['_on_sale_checkbox_field'] ) ) checked( $checkbox_stored_meta['_on_sale_checkbox_field'][0], 'yes' ); ?>>
         <span class="text-sm">
-          <?php echo __('Is on sale?', 'twentytwenty_child') ?>
+          <?php echo __('IS ON SALE ?', 'twentytwenty_child') ?>
         </span>
       </label>
     </div>
 
   </div>
 
-
+<!-- Product image gallery (6 images) -->
   <div id="custom-images">
 
     <div class="custom-img-container flex flex-wrap -mx-3 my-6">
@@ -109,7 +116,7 @@ function products_add_fields(){
           <div class="flex flex-wrap my-3">
             <a class="delete-custom-img bg-red-500 hover:bg-red-700 text-white hover:text-white font-bold py-2 px-4 rounded"
               href="#">
-              <?php echo __('Remove this image', 'twentytwenty_child'); ?>
+              <?php echo __('REMOVE THIS IMAGE', 'twentytwenty_child'); ?>
             </a>
           </div>
         </div>
@@ -125,11 +132,12 @@ function products_add_fields(){
 
     <a class="upload-custom-img bg-blue-500 hover:bg-blue-700 text-white hover:text-white font-bold py-2 px-4 rounded <?php if ( $you_have_img  ) { echo 'hidden'; } ?>"
       href="<?php echo $upload_link; ?>">
-      <?php echo __('Add custom image', 'twentytwenty_child'); ?>
+      <?php echo __('ADD GALLERY IMAGE', 'twentytwenty_child'); ?>
     </a>
-
+    <p class="text-gray-600 mt-2 w-full text-xs italic">Add images to product gallery</p>
   </div>
 
+  <!-- Description -->
   <div class="flex flex-wrap -mx-3 mb-6">
     <div class="w-full px-3">
 
@@ -139,22 +147,23 @@ function products_add_fields(){
       <textarea style="resize: none;" rows="5"
         class="appearance-none block w-full bg-gray-200 text-gray-700 border rounded py-3 px-4 mb-3 leading-tight focus:outline-none focus:bg-white"
         name="_description_textarea_field"
-        placeholder="Description"><?php echo $description_textareafield; ?></textarea>
+        placeholder="<?php echo __('Description', 'twentytwenty_child'); ?>"><?php echo $description_textareafield; ?></textarea>
 
     </div>
   </div>
 
 
-
+<!-- YouTube video -->
   <div class="flex flex-wrap -mx-3 mb-6">
     <div class="w-full px-3">
       <label class="block uppercase tracking-wide text-gray-700 text-xs font-bold mb-2" for="grid-last-name">
-        <?php echo __('YouTube video', 'twentytwenty_child') ?>
+        <?php echo __('YouTube video', 'twentytwenty_child'); ?>
       </label>
       <input
         class="appearance-none block w-full bg-gray-200 text-gray-700 border border-gray-200 rounded py-3 px-4 leading-tight focus:outline-none focus:bg-white focus:border-gray-500"
         type="text" name="_youtube_video_text_field" value="<?php echo $youtube_video_textfield; ?>"
-        placeholder="YouTube video">
+        placeholder="<?php echo __('YouTube video', 'twentytwenty_child'); ?>">
+        <p class="text-gray-600 mt-2 w-full text-xs italic">Add YouTube url to field</p>
     </div>
   </div>
 
@@ -208,9 +217,11 @@ else {
       $meta_value = get_post_meta( $post->ID, 'image_src', false );
       
     /* For looping all meta values */
+    if($meta_value){
       foreach ($meta_value as $value){
         delete_post_meta( $post->ID, 'image_src', $value );
       }
+    }
       
     /* Get the posted data and sanitize it for use as an HTML class. */
     if(isset($_POST["image_src"])) {
